OpenSCAD and stl files for the C-Arm Phantom Stand
The C-Arm phantom stand is the largest of the 3D printed testing aids.  This requires a printer such as the Makerbot Z18.  The stand is composed of a top piece which holds aluminum blocks, Copper sheets and lead beam blockers and has an insert that will hold the R/F sensor from Raysafe.  There is a single leg file, 4 of which are needed.  The leg is available in straight or spiral designs.  There are two bases, one for image intensifiers in the 9 inch or larger size range, and a small base for miniature c-arms.