Film holder for collimation testing
This file contains the OpenSCAD and stl files to make the radiochromic film strip holder and the track to slide it in.  To complete this device, cut three pieces of wire (we used paper clips) that are slightly narrower than the film holder, and glue them in the notches in the top of the film holder.  These notches are set at 2 cm spacing.  To provide friction to prevent the film strip holder from falling out of the track, we used a narrow strip of the fuzzy side of hook and loop fastener material.  A small strip placed on each side of the track under the overhanging lip will provide enough friction to keep the film holder in place but still allow it to move easily enough for adjustment