Radiographic/Fluoroscopic Phantom Stand
This phantom stand is designed for use in R/F rooms with under-table tubes.  The purpose of this stand is to allow the x-ray detector to be placed under the aluminum phantom blocks.  The x-ray detector holder is integral to this stand and is designed for the Raysafe X2 R/F sensor.