Raysafe stand
This is a stand designed for the Raysafe X2 base unit.  This should also work with the Raysafe X2 Solo and may work with other dosimetry system base units of a similar size.  The purpose is to angle the display so that it is easily visible from across a room.  We use it mostly for CT surveys.